% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_m.R
\name{generate_m_list}
\alias{generate_m_list}
\title{These functions are dedicated to using the validated version of the
transition probabilities which the functions in \code{validate_tp.R} produce to
generate the sparce matrices required to extrapolate the sequencing model}
\usage{
generate_m_list(
  m_specification,
  transition_prob_list,
  first_state_name = "initial"
)
}
\arguments{
\item{m_specification}{A list containing model parameters, including:
\itemize{
\item n_cycles: Number of model cycles.
\item pre_tunnels: Number of pre-tunnel states.
\item matrix_size: Size of the transition matrix.
\item m1_ijx: Coordinate matrix for m1.
\item m2_ijx: Coordinate matrix for m2.
}}

\item{transition_prob_list}{A list of transition probabilities for each
state.}
}
\value{
A list with two elements:
\itemize{
\item m1: A list of sparse matrices (one per cycle) for pre-tunnel states.
\item m2: A sparse matrix for tunnel states.
}
}
\description{
See the vignette \code{sparse-matrix-functions.Rmd} for more details on how these
work.
Generate Markov Transition Matrices (m1 and m2) Based on Model Specification
}
\details{
This function generates two transition matrices (m1 and m2) for a Markov
model, using the provided model specification and a validated list of
transition probabilities.
\itemize{
\item m1 corresponds to pre-tunnel states for each model cycle.
\item m2 corresponds to tunnel states after entering the sequence.
}

The function validates the transition probability list, splits it into
pre-tunnel and tunnel components, and constructs the corresponding sparse
matrices. It ensures that row sums are appropriate and asserts that the
tunnel matrix rows sum to 1.
}
